+++ src/parser/float_expr_ast.cpp
@@ -29,7 +29,7 @@ std::ostream& Parser::AST::operator<< (std::ostream& stream, const t_float_expr_
 void Parser::AST::OperandPrinterVisitor::operator()(const t_float_expr_nil &nil) const {
     m_out << "<float_expr_operand value=\"null\"></float_expr_operand>";
 }
-void Parser::AST::OperandPrinterVisitor::operator()(const ::boost::spirit::x3::variant<float, std::string> &v) const {
+void Parser::AST::OperandPrinterVisitor::operator()(const ::boost::spirit::x3::variant<double, std::string> &v) const {
     m_out << "<float_expr_operand value=\"";
     ::boost::apply_visitor(Parser::AST::TFloatPrinterVisitor(m_out), v);
     m_out << "\"></float_expr_operand>";
@@ -41,7 +41,7 @@ void Parser::AST::OperandPrinterVisitor::operator()(const ::boost::spirit::x3::f
     m_out << ast.get();
 }
 
-void Parser::AST::TFloatPrinterVisitor::operator()(const float &f) const {
+void Parser::AST::TFloatPrinterVisitor::operator()(const double &f) const {
     m_out << f;
 }
